% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{political_ideology}
\alias{political_ideology}
\title{Political Ideology Dataset}
\format{
A data.table with 250 rows and 5 variables:
\describe{
  \item{pmid}{Character: PubMed ID}
  \item{year}{Character: Year of publication}
  \item{journal}{Character: Name of the journal}
  \item{articletitle}{Character: Title of the article}
  \item{abstract}{Character: Abstract of the article}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/}
}
\usage{
political_ideology
}
\description{
A dataset containing information on political ideology articles from PubMed.
}
\keyword{datasets}
