% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prompts}
\alias{prompts}
\title{Prompts Data Set}
\format{
A named list where each element is a character string representing the content of a text file. The names of the elements correspond to the file names without the .txt extension.
\describe{
  \item{\code{prompt1}}{The content of the first text file.}
  \item{\code{prompt2}}{The content of the second text file.}
  ...
}
}
\source{
The text files were obtained from the directory `/home/jtimm/pCloudDrive/GitHub/packages/prompts/prompts`.
}
\usage{
prompts
}
\description{
This data set contains a list of character strings, each representing the content of a text file.
The text files are prompts that can be used for various purposes, such as testing and examples.
}
\examples{
data(prompts)
print(prompts[["prompt1"]])

}
\keyword{datasets}
