% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine Sample Dataset}
\format{
A data frame with 250 rowine and the following columns:
\describe{
  \item{X}{Index of the wine}
  \item{country}{Country of origin}
  \item{description}{Description of the wine}
  \item{designation}{Designation of the wine}
  \item{points}{Points awarded to the wine}
  \item{price}{Price of the wine}
  \item{province}{Province where the wine is produced}
  \item{region_1}{Primary region where the wine is produced}
  \item{variety}{Variety of the wine}
  \item{winery}{Winery that produces the wine}
}
}
\usage{
wine
}
\description{
This dataset contains a sample of 250 wines with various attributes.
}
\examples{
data(wine)
head(wine)
}
\keyword{datasets}
