% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\docType{data}
\name{prompts}
\alias{prompts}
\title{Some example prompts for demo and testing purposes}
\format{
An object of class \code{list} of length 3.
}
\usage{
prompts
}
\description{
Some example prompts for demo and testing purposes
}
\keyword{datasets}
